+++ cextern/wcslib/C/flexed/wcsulex.c
@@ -7151,7 +7151,7 @@ static const yy_state_type yy_NUL_trans[375] =
 #line 1 "wcsulex.l"
 /*============================================================================
 
-  WCSLIB 7.2 - an implementation of the FITS WCS standard.
+  WCSLIB 7.3 - an implementation of the FITS WCS standard.
   Copyright (C) 1995-2020, Mark Calabretta
 
   This file is part of WCSLIB.
@@ -7173,7 +7173,7 @@ static const yy_state_type yy_NUL_trans[375] =
 
   Author: Mark Calabretta, Australia Telescope National Facility, CSIRO.
   http://www.atnf.csiro.au/people/Mark.Calabretta
-  $Id: wcsulex.c,v 7.2 2020/03/09 07:31:23 mcalabre Exp $
+  $Id: wcsulex.c,v 7.3 2020/06/03 03:37:02 mcalabre Exp $
 *=============================================================================
 *
 * wcsulex.l is a Flex description file containing the definition of a